backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
python-multipart
websockets
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic
pydantic-settings
